GetItemSlotName should only use g_slot_cache if class is TFClass_Unknown
based on a true story

TF2IDB_GetItemSlot(414, TFClass_Pyro);
"hello plugin, in what slot does the Pyro wear the Liberty Launcher"

"that is a good question, let me first run g_statement_GetItemSlotNameByClass which correctly checks for the item id and class, but because there is obviously not going to be a result I will contact g_slot_cache instead which just checks for the item id, it goes into primary slot by the way!"

yeah, no. what about we actually make it so GetItemSlotName actually returns false if the item/class combination is not valid.